# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SeniorShield API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voice/process-request:
    post:
      operationId: processVoiceRequest
      tags: [voice]
      summary: Process a voice request (text input version)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceTextRequest"
      responses:
        "200":
          description: Voice response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceResponse"
  /voice/history:
    get:
      operationId: getVoiceHistory
      tags: [voice]
      summary: Get voice assistance history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Voice history
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: "#/components/schemas/VoiceHistoryItem"
                required:
                  - history
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VoiceTextRequest:
      type: object
      properties:
        request_text:
          type: string
      required:
        - request_text
    VoiceResponse:
      type: object
      properties:
        request_text:
          type: string
        response_text:
          type: string
        task_category:
          type: string
        success:
          type: boolean
        error_message:
          type: string
        history_id:
          type: string
      required:
        - request_text
        - response_text
        - task_category
        - success
    VoiceHistoryItem:
      type: object
      properties:
        id:
          type: string
        request_text:
          type: string
        response_text:
          type: string
        task_category:
          type: string
        success:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - request_text
        - response_text
        - task_category
        - success
        - created_at
...